[26] - Uniformidad directorios en app/Application
