+++ app/Infrastructure/Controllers/GetWalletCryptocurrenciesController.php
@@ -3,7 +3,7 @@
 namespace App\Infrastructure\Controllers;
 
 use App\Application\Exceptions\WalletNotFoundException;
-use App\Application\WalletCryptocurrenciesService;
+use App\Application\WalletCryptocurrencies\WalletCryptocurrenciesService;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Response;
 use Illuminate\Routing\Controller as BaseController;
