+++ tests/app/Application/SellCoinServiceTest.php
@@ -4,7 +4,7 @@ namespace Tests\app\Application;
 
 use App\Application\Exceptions\CoinNotFoundException;
 use App\Application\Exceptions\WalletNotFoundException;
-use App\Application\SellCoinService;
+use App\Application\SellCoin\SellCoinService;
 use App\Application\WalletDataSource\WalletDataSource;
 use App\Domain\Coin;
 use App\Domain\Wallet;
